Last fixes
