+++ src/components/gameCanvas/GameCanvas.tsx
@@ -3,8 +3,17 @@ import { useSelector, useDispatch } from 'react-redux';
 import cn from 'classnames';
 
 import GameMain from '../../game/gamemain';
-import { setIsGameStarted, setIsGamePaused, setIsGameOver, setLastScore } from 'store/game/actions';
-import { gameIsGameStartedSelector, gameIsGamePausedSelector, gameIsGameOverSelector } from 'store/game/selectors';
+import {
+  setIsGameStarted,
+  setIsGamePaused,
+  setIsGameOver,
+  setLastScore
+} from 'store/game/actions';
+import {
+  gameIsGameStartedSelector,
+  gameIsGamePausedSelector,
+  gameIsGameOverSelector
+} from 'store/game/selectors';
 import { MENU_ACTIONS } from 'constants/menuConstants';
 import { GAME_OPTIONS } from 'constants/gameConstants';
